refactor(web): ロール判定ロジックをuseRoleフックに集約
ROLE_LEVEL・ROLE_LABELSの定義とロール判定ロジックを
各コンポーネントからuseRoleフックに抽出し、一箇所で管理するよう変更。
